+++ plugins/modules/nspartition.py
@@ -29,6 +29,7 @@ options:
       - present
       - absent
       - switched
+      - unset
     default: present
     description:
       - The state of the resource being configured by the module on the NetScaler
@@ -37,6 +38,7 @@ options:
         the module's parameters.
       - When C(absent), the resource will be deleted from the NetScaler ADC node.
       - When C(switched), the partition will be switched on the NetScaler ADC node.
+      - When C(unset), the resource will be unset on the NetScaler ADC node.
     type: str
   force:
     type: bool